Add settings for selecting when desktop icons should be hidden
Only the GUI is implemented, it doesn't yet work. And some things that are
planned for later have also been added to the GUI.

When making this, I sourced knowledge from my previous Python OBS scripting work [1],
example plugins written in C# [2] and Go [3]. And of course the OBS documentation,
but it wasn't that useful.

[1]: https://github.com/Susko3/obs-youtube-live-qa
[2]: https://github.com/YorVeX/ObsCSharpExample
[3]: https://github.com/fzwoch/obs-teleport